requirements.txt
deepdiff==5.2.3
Django==3.1.7
djangorestframework==3.12.2
django-allauth==0.44.0
django-cryptography==1.0
django-filter==2.4.0
gunicorn==20.0.4
praw==7.2.0
psycopg2-binary==2.8.6
pydantic==1.8.1
requests==2.25.1
